Since full.core specifies :aot all there are issues
This is because all of its libraries it depends on are packaged into the
jar that is published as a library. One of these libraries, ns-tracker
depends org.clojure/tools.namespace 0.2.11. Later versions of midje
are dependeny on or.clojure/tools.namespace 0.3.0. This causes issues
for libraries that try to include full.core and latest version of midje.

To fix this we force the version of tools.namespace to 0.3.0 using
managed dependencies section